Display the packages with wheels compatible with the newest Python
Motivation:
Every year, the Fedora Python contributors open a bunch of issues to
upstream projects to ask them to publish the new Python wheels early.
This update will increase the visibility of the projects that need
our attention and may serve as a gentle nudge for the project authors
to not stay behind the rest of the ecosystem.